% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudtrail_operations.R
\name{list_public_keys}
\alias{list_public_keys}
\title{Returns all public keys whose private keys were used to sign the digest files within the specified time range}
\usage{
list_public_keys(StartTime = NULL, EndTime = NULL, NextToken = NULL)
}
\arguments{
\item{StartTime}{Optionally specifies, in UTC, the start of the time range to look up public keys for CloudTrail digest files. If not specified, the current time is used, and the current public key is returned.}

\item{EndTime}{Optionally specifies, in UTC, the end of the time range to look up public keys for CloudTrail digest files. If not specified, the current time is used.}

\item{NextToken}{Reserved for future use.}
}
\description{
Returns all public keys whose private keys were used to sign the digest files within the specified time range. The public key is needed to validate digest files that were signed with its corresponding private key.
}
\details{
CloudTrail uses different private/public key pairs per region. Each digest file is signed with a private key unique to its region. Therefore, when you validate a digest file from a particular region, you must look in the same region for its corresponding public key.
}
\section{Accepted Parameters}{
\preformatted{list_public_keys(
  StartTime = as.POSIXct("2015-01-01"),
  EndTime = as.POSIXct("2015-01-01"),
  NextToken = "string"
)
}
}

