% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudtrail_operations.R
\name{list_tags}
\alias{list_tags}
\title{Lists the tags for the trail in the current region}
\usage{
list_tags(ResourceIdList, NextToken = NULL)
}
\arguments{
\item{ResourceIdList}{[required] Specifies a list of trail ARNs whose tags will be listed. The list has a limit of 20 ARNs. The format of a trail ARN is:

\code{arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail}}

\item{NextToken}{Reserved for future use.}
}
\description{
Lists the tags for the trail in the current region.
}
\section{Accepted Parameters}{
\preformatted{list_tags(
  ResourceIdList = list(
    "string"
  ),
  NextToken = "string"
)
}
}

