% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudtrail_package.R
\docType{package}
\name{paws.cloudtrail-package}
\alias{paws.cloudtrail}
\alias{paws.cloudtrail-package}
\title{paws.cloudtrail: AWS CloudTrail}
\description{
This is the CloudTrail API Reference. It provides
descriptions of actions, data types, common parameters, and common
errors for CloudTrail.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
