% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudtrail_operations.R
\name{remove_tags}
\alias{remove_tags}
\title{Removes the specified tags from a trail}
\usage{
remove_tags(
  ResourceId = "string",
  TagsList = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
\arguments{
\item{ResourceId}{[required] Specifies the ARN of the trail from which tags should be removed. The format of a trail ARN is:

\code{arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail}}

\item{TagsList}{Specifies a list of tags to be removed.}
}
\description{
Removes the specified tags from a trail.
}
\examples{

}
